tests: rename test_annotate_command.rs to test_file_annotate_command.rs
