+++ cli/tests/runner.rs
@@ -12,7 +12,6 @@ mod test_abandon_command;
 mod test_acls;
 mod test_advance_bookmarks;
 mod test_alias;
-mod test_annotate_command;
 mod test_backout_command;
 mod test_bookmark_command;
 mod test_builtin_aliases;
@@ -29,6 +28,7 @@ mod test_diffedit_command;
 mod test_duplicate_command;
 mod test_edit_command;
 mod test_evolog_command;
+mod test_file_annotate_command;
 mod test_file_chmod_command;
 mod test_file_show_command;
 mod test_file_track_untrack_commands;
