+++ cli/tests/test_file_annotate_command.rs
